Vector Graphics game
Fly a space ship around and destroy enemies before they run into you!

This was my first project in Java, which is why the source directory is
"Hello World". I followed a tutorial on youtube to get started but it
became plain that the author of the tutorial was not a very good
programmer. That is why you will see fragments in my game loop that were
designed to set up parallel processing. Future projects are more
concise, given what I learned in this one.